uid 타입을 bigint에서 int로 변경

+++ config/init.mariadb.sql
@@ -6,7 +6,7 @@ create sequence seq_arc_response_uid start with 1 increment by 1;
 --
 create table thread
 (
-    thread_uid bigint(0) unsigned not null,
+    thread_uid int(0) unsigned not null,
     board_uid    varchar(10)        not null,
     title    varchar(50)        not null,
     password varchar(256)       not null,
@@ -20,8 +20,8 @@ create table thread
 --
 create table response
 (
-    response_uid    bigint(0) unsigned not null,
-    thread_uid    bigint(0) unsigned not null,
+    response_uid    int(0) unsigned not null,
+    thread_uid    int(0) unsigned not null,
     sequence      int(0) unsigned    not null,
     user_name   varchar(60)        not null,
     user_id     varchar(10)        not null,
@@ -43,9 +43,9 @@ create table response
 --
 create table arc_response
 (
-    arc_response_uid bigint(0) unsigned not null,
-    response_uid    bigint(0) unsigned not null,
-    thread_uid    bigint(0) unsigned not null,
+    arc_response_uid int(0) unsigned not null,
+    response_uid    int(0) unsigned not null,
+    thread_uid    int(0) unsigned not null,
     sequence      int(0) unsigned    not null,
     user_name   varchar(60)        not null,
     user_id     varchar(10)        not null,
