+++ config/init.mysql.sql
@@ -1,12 +1,12 @@
-create table seq_thread_uid (sequence bigint(0) auto_increment primary key) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
+create table seq_thread_uid (sequence int(0) auto_increment primary key) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
 --
-create table seq_response_uid (sequence bigint(0) auto_increment primary key) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
+create table seq_response_uid (sequence int(0) auto_increment primary key) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
 --
-create table seq_arc_response_uid (sequence bigint(0) auto_increment primary key) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
+create table seq_arc_response_uid (sequence int(0) auto_increment primary key) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
 --
 create table thread
 (
-    thread_uid bigint(0) unsigned not null,
+    thread_uid int(0) unsigned not null,
     board_uid    varchar(10)        not null,
     title    varchar(50)        not null,
     password varchar(256)       not null,
@@ -20,8 +20,8 @@ create table thread
 --
 create table response
 (
-    response_uid    bigint(0) unsigned not null,
-    thread_uid    bigint(0) unsigned not null,
+    response_uid    int(0) unsigned not null,
+    thread_uid    int(0) unsigned not null,
     sequence      int(0) unsigned    not null,
     user_name   varchar(60)        not null,
     user_id     varchar(10)        not null,
@@ -43,9 +43,9 @@ create table response
 --
 create table arc_response
 (
-    arc_response_uid bigint(0) unsigned not null,
-    response_uid    bigint(0) unsigned not null,
-    thread_uid    bigint(0) unsigned not null,
+    arc_response_uid int(0) unsigned not null,
+    response_uid    int(0) unsigned not null,
+    thread_uid    int(0) unsigned not null,
     sequence      int(0) unsigned    not null,
     user_name   varchar(60)        not null,
     user_id     varchar(10)        not null,
